@@ -30,6 +30,7 @@
     "express": "^4.21.2",
     "form-urlencoded": "3.0.0",
     "http-errors": "~1.6.3",
+    "inquirer": "^12.10.0",
     "jquery": "^3.5.1",
     "jsonwebtoken": "^9.0.0",
     "jwks-rsa": "2.1.4",
@@ -55,7 +56,7 @@
     "office-addin-test-helpers": "^2.0.3",
     "rimraf": "^6.0.1",
     "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "typescript": "^5.6.0"
   },
   "repository": {
     "type": "git",
